Week8

+++ homeworks/week8/hw2/deleteCommentsChild.php
@@ -1,23 +1,26 @@
 <?php
+  session_start();
   require_once('conn.php');
  
-      
+  //判斷是否登入過
+  if(isset($_SESSION["member_id"])) {
+    $stmt = $conn->prepare("SELECT * FROM annieshen_users WHERE user_id=?");
+    $stmt->bind_param("s", $_SESSION["member_id"]);
+    $stmt->execute();
+    $result = $stmt->get_result();
+    $row = $result->fetch_assoc();
+    $user_id=$row['user_id'];
+  }   
 
   // 刪除子留言
   if(isset($_POST['comment_id'])){
       $comment_id = $_POST['comment_id'];
-      
-      $sql = "SELECT * FROM annieshen_commentschild WHERE id = ?";
-      $stmt = $conn->prepare($sql);
-      $stmt->bind_param("i", $comment_id);
-      $stmt->execute();
-      $stmt->store_result();
-      $row = $stmt->num_rows;
-      if($row > 0){
-          $deleteSql = "DELETE FROM annieshen_commentschild WHERE id = ?";
-          $stmt_deleteSql = $conn->prepare($deleteSql);
-          $stmt_deleteSql->bind_param("i", $comment_id);
-          $stmt_deleteSql->execute();
+  
+      $deleteSql = "DELETE FROM annieshen_commentschild WHERE id = ? AND user_id = ?";
+      $stmt_deleteSql = $conn->prepare($deleteSql);
+      $stmt_deleteSql->bind_param("is", $comment_id, $user_id);
+      $result = $stmt_deleteSql->execute();
+      if($result){
           header('Location: index.php');
       }else{
         echo '失敗';
